wroet unit tests for LexiconHelpers.js
