+++ src/js/helpers/ResourcesHelpers.js
@@ -5,7 +5,7 @@ import ospath from 'ospath';
 // helpers
 import * as BibleHelpers from './bibleHelpers';
 // constants
-export const USER_RESOURCES_PATH = path.join(ospath.home(), 'translationCore/resources');
+export const USER_RESOURCES_PATH = path.join(ospath.home(), 'translationCore', 'resources');
 export const STATIC_RESOURCES_PATH = path.join(__dirname, '../../../tC_resources/resources');
 
 /**
@@ -262,25 +262,4 @@ export function getGLQuote(languageId, groupId, currentToolName) {
     const resourceIndexArray = fs.readJSONSync(GLQuotePathIndex);
     return resourceIndexArray.find(({ id }) => id === groupId).name;
   } catch (e) { return null }
-}
-
-export function getLexiconData(lexiconId, entryId) {
-  try {
-    let languageId = 'en';
-    let resourceVersion = 'v0';
-    // generate path from resourceType and articleId
-    let lexiconPath = path.join(USER_RESOURCES_PATH, languageId, 'lexicons', lexiconId, resourceVersion, 'content');
-    let entryPath = path.join(lexiconPath, entryId + '.json');
-    let entryData;
-    if (fs.existsSync(entryPath)) {
-      entryData = fs.readJsonSync(entryPath, 'utf8'); // get file from fs
-    }
-    return {
-      [lexiconId]: {
-        [entryId]: entryData
-      }
-    };
-  } catch (error) {
-    console.error(error);
-  }
 }
